Conflicts:
	bitrix/templates/estelife/components/bitrix/news.list/ajax/result_modifier.php

+++ bitrix/templates/estelife/components/bitrix/news.list/ajax/result_modifier.php
@@ -16,7 +16,11 @@ if(isset($arResult['ITEMS'])){
 	$arAvai=array('SRC','NAME','PREVIEW_TEXT','DETAIL_PAGE_URL','ACTIVE_FROM');
 
 	foreach($arResult['ITEMS'] as &$arItem){
-		$arItem['PREVIEW_TEXT']=\core\types\VString::truncate($arItem['PREVIEW_TEXT'], 70, '...').'<span></span>';
+		if (!empty($arItem['PROPERTIES']['SHORT_TEXT']['VALUE']['TEXT'])){
+			$arItem['PREVIEW_TEXT']=$arItem['PROPERTIES']['SHORT_TEXT']['VALUE']['TEXT'].'<span></span>';
+		}else{
+			$arItem['PREVIEW_TEXT']=\core\types\VString::truncate($arItem['PREVIEW_TEXT'], 70, '...').'<span></span>';
+		}
 
 		if(!empty($arItem['ACTIVE_FROM']))
 			$arItem['ACTIVE_FROM']=date('d.m.Y',strtotime($arItem['ACTIVE_FROM']));
